% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeightsInImage.R
\name{AnnotationHeight}
\alias{AnnotationHeight}
\title{Calculate height of annotation in image}
\usage{
AnnotationHeight(X1, Y1, X2, Y2, CamSetup)
}
\arguments{
\item{X1}{x pixel coordinate of base of annotation object}

\item{Y1}{y pixel coordinate of base of annotation object}

\item{X2}{x pixel coordinate of top of annotation object}

\item{Y2}{y pixel coordinate of top of annotation object}

\item{CamSetup}{details of camera as output from function CamSetup}
}
\value{
A list with the following elements
X1: x pixel coordinate of base of annotation object
Y1: y pixel coordinate of base of annotation object
X2: x pixel coordinate of top of annotation object
Y2: y pixel coordinate of top of annotation object
NadirX: x pixel coordinate of point directly beneath camera
NadirY: y pixel coordinate of point directly beneath camera
PPX: x pixel denoting principle point of image (middle)
PPY: y pixel denoting principle point of image (middle)
CamConstantC: camera constant C
CamHeight: height of camera above the ground (mm)
ObjPixX: width of annotation in pixels
ObjPixY: height of annotation in pixels
ObjPixXY: length of annotation in pixels
ObjAngR: angle of annotation line in the image
A: intermediate calculation A
B: intermediate calculation B
C: intermediate calculation C
D: intermediate calculation D
E: intermediate calculation E
Eq1: equation 1 (Verykokou & Ioannidis 2015)
Eq2: equation 2 (Verykokou & Ioannidis 2015)
Eq3: equation 3 (Verykokou & Ioannidis 2015)
Eq4: equation 4 (Verykokou & Ioannidis 2015)
ObjHeight: Height of object above the ground (mm)
ObjWidth: width of object as seen in image (mm)
ObjLength: Length of object as seen in image (mm - assumes object is inclined on a plane perpendicular to the direction the camera is pointing)
}
\description{
Calculate height of annotation in image
}
\details{
Calculate height of an annotation in an image where camera specifics such as FOV, height & angle are known
}
\examples{
myCamSetup <- CamSetup(122.6, 94.4, 94.4, 55, 2704, 1520, 6.17, 4.65, 28.8, 550)
AnnotationHeight(100, 100, 100, 50, myCamSetup)
}
