% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeightsInImage.R
\name{CamSetup}
\alias{CamSetup}
\title{Calculate camera nadir and constant C}
\usage{
CamSetup(
  FullHFOV,
  FullVFOV,
  UsedHFOV,
  UsedVFOV,
  PixDimX,
  PixDimY,
  FullSensorWidth,
  FullSensorHeight,
  CamAngle,
  CamHeight
)
}
\arguments{
\item{FullHFOV}{Value in degrees of the horizontal field of view of a camera when using the full sensor (e.g. 122.6 for gopro 5 in 4:3 W mode)}

\item{FullVFOV}{Value in degrees of the vertical field of view of a camera when using the full sensor (e.g. 94.4 for gopro 5 in 4:3 W mode)}

\item{UsedHFOV}{Value in degrees of the horizontal field of view of the camera as used in the image (e.g. 94.4 for gopro 5 in 16:9 M mode)}

\item{UsedVFOV}{Value in degrees of the vertical field of view of the camera as used in the image (e.g. 55.0 for gopro 5 in 16:9 W mode)}

\item{PixDimX}{Pixel dimension of image in X dimension (image width)}

\item{PixDimY}{Pixel dimension of image in Y dimension (image height)}

\item{FullSensorWidth}{width of sensor in mm as reported by manufacturer (e.g. gopro 5 has width 6.17)}

\item{FullSensorHeight}{height of sensor in mm as reported by manufacturer (e.g. gopro 5 has width 4.65)}

\item{CamAngle}{Camera angle in degrees relative to the horizon (0=horizontal, 90=downward facing)}

\item{CamHeight}{Camera height in mm above the ground}
}
\value{
A list with the following elements
FullHFOV: Value in degrees of the horizontal field of view of a camera when using the full sensor
FullVFOV: Value in degrees of the vertical field of view of a camera when using the full sensor
UsedHFOV: Value in degrees of the horizontal field of view of the camera as used in the image
UsedVFOV: Value in degrees of the vertical field of view of the camera as used in the image
FullSensorWidth: width of sensor in mm as reported by manufacturer
FullSensorHeight: height of sensor in mm as reported by manufacturer
UsedSensorWidth: width of used part of the sensor in mm
CamAngle: Angle of camera
CamHeight: Height of camera above the ground (mm)
FocalLength: Calculated focal length of the camera
CamConstantC: Calculated camera constant C
PixDimX:  pixel dimension of image on x-axis (image width)
PixDimY:  pixel dimension of image on y-axis (image height)
PPX:      principle point of the image on the X axis (middle of image)
PPX:      principle point of the image on the Y axis (middle of image)
NadirX:   X pixel coordinate of point directly beneath camera
NadirY:   Y pixel coordinate of point directly beneath camera
}
\description{
Calculate camera nadir and constant C
}
\details{
call functions NadirXY and CamConstantC and package output into a single list object
}
\examples{
myCamSetup <- CamSetup(122.6, 94.4, 94.4, 55, 2704, 1520, 6.17, 4.65, 28.8, 550)
}
