% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeightsInImage.R
\name{getCamConstantC}
\alias{getCamConstantC}
\title{Calculate the camera constant C}
\usage{
getCamConstantC(
  FullHFOV,
  FullVFOV,
  UsedHFOV,
  UsedVFOV,
  PixDimX,
  PixDimY,
  FullSensorWidth,
  FullSensorHeight
)
}
\arguments{
\item{FullHFOV}{Value in degrees of the horizontal field of view of a camera when using the full sensor (e.g. 122.6 for gopro 5 in 4:3 W mode)}

\item{FullVFOV}{Value in degrees of the vertical field of view of a camera when using the full sensor (e.g. 94.4 for gopro 5 in 4:3 W mode)}

\item{UsedHFOV}{Value in degrees of the horizontal field of view of the camera as used in the image (e.g. 94.4 for gopro 5 in 16:9 M mode)}

\item{UsedVFOV}{Value in degrees of the vertical field of view of the camera as used in the image (e.g. 55.0 for gopro 5 in 16:9 W mode)}

\item{PixDimX}{Pixel dimension of image in X dimension (image width)}

\item{PixDimY}{Pixel dimension of image in Y dimension (image height)}

\item{FullSensorWidth}{width of sensor in mm as reported by manufacturer (e.g. gopro 5 has width 6.17)}

\item{FullSensorHeight}{height of sensor in mm as reported by manufacturer (e.g. gopro 5 has width 4.65)}
}
\value{
A list containing the following elements
FullHFOV: Value in degrees of the horizontal field of view of a camera when using the full sensor
FullVFOV: Value in degrees of the vertical field of view of a camera when using the full sensor
UsedHFOV: Value in degrees of the horizontal field of view of the camera as used in the image
UsedVFOV: Value in degrees of the vertical field of view of the camera as used in the image
FullSensorWidth: width of sensor in mm as reported by manufacturer
FullSensorHeight: height of sensor in mm as reported by manufacturer
PixDimX:  pixel dimension of image on x-axis (image width)
PixDimY:  pixel dimension of image on y-axis (image height)
LenAF:    length (mm) between the lens and the sensor
AngACF:   Angle equal to half full FOV
AngCAF:   Angle = 90-AngACF
FocalLength: Calculated focal length of the camera
AngECF:   Angle equal to half the Used FOV
AngCEF:   Angle = 90-AngCEF
LenEF:    Length (mm) half the used sensor width
UsedSensorWidth: width of used part of the sensor in mm
PixSize: UsedSensorWidth / PixDimX
CamConstantC: Calculated camera constant C
}
\description{
Calculate the camera constant C
}
\details{
Camera constant C is calculated based on the sensor size, pixel dimensions of the image
We are using formula c = f / p where f = focal length and p = pixel size
where p = X / W and X = pixel dimension of image and W = width of sensor
and where f is derived trigonimcally based on a triangle of adjacent length = half of sensor width and adjacent angle = 180 - (FOV/2)
}
\examples{
cc<-getCamConstantC(122.6, 94.4, 94.4, 55.0, 2704, 1520, 6.17, 4.65)
}
