% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateArea.R
\name{ImageArea}
\alias{ImageArea}
\title{Estimate area seen in tilted camera pointing at the seabed}
\usage{
ImageArea(Height, Angle, VFOV, HFOV, Proportion = 0.99)
}
\arguments{
\item{Height}{\itemize{
\item Height of the camera above the seabed (in m)
}}

\item{Angle}{\itemize{
\item Angle (in dgrees) of the camera pointing down from the horizontal
}}

\item{VFOV}{\itemize{
\item Vertical field of view angle (in degrees)
}}

\item{HFOV}{\itemize{
\item Horizontal field of view angle (in degrees)
}}

\item{Proportion}{\itemize{
\item Proportion of image to use to estimate area (0-1 values default = 0.99)
}}
}
\value{
List containing the following calculations used in estimating the area (see Long et al 2020):
Height - input parameter
Angle - input parameter
VFOV - input parameter
HFOV - input paramter
Alpha - Angle (radians)
Beta - Angle (radians)
Delta - Angle (radians)
Theta - Angle (radians)
LengthAE - Length (m)
LengthBD - Length (m)
LengthGF - Length (m)
AreaS - Area (m2)
ImageArea - Area (m2)
}
\description{
Estimate area seen in tilted camera pointing at the seabed
}
\details{
Estimate area seen in tilted camera pointing at the seabed based on Nakajima, R. et al. (2014) A new method for estimating the area of the seafloor from oblique images taken by deep-sea submersible survey platforms. JAMSTEC Report of Research and Development, 19, pp.59-66, as used in Long et al (2020) https://www.frontiersin.org/articles/10.3389/fmars.2020.00460/full
}
\examples{
x <- ImageArea(0.55, 28.8, 40.3, 66.4)
}
