% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateArea.R
\name{ImageWidth}
\alias{ImageWidth}
\title{Estimate width of image at a given point seen in tilted camera pointing at the seabed}
\usage{
ImageWidth(Height, Angle, VFOV, HFOV, Position = 0.5)
}
\arguments{
\item{Height}{\itemize{
\item Height of the camera above the seabed (in m)
}}

\item{Angle}{\itemize{
\item Angle (in dgrees) of the camera pointing down from the horizontal
}}

\item{VFOV}{\itemize{
\item Vertical field of view angle (in degrees)
}}

\item{HFOV}{\itemize{
\item Horizontal field of view angle (in degrees)
}}

\item{Position}{\itemize{
\item Position in the image to calculate width (Proption 0-1, where 0 is along the bottom of the image, 1 is the top and 0.5 (default) is the midline.
}}
}
\value{
width of the image in m at the position selected
}
\description{
Estimate width of image at a given point seen in tilted camera pointing at the seabed
}
\details{
Estimate image width seen in tilted camera pointing at the seabed based on Nakajima, R. et al. (2014) A new method for estimating the area of the seafloor from oblique images taken by deep-sea submersible survey platforms. JAMSTEC Report of Research and Development, 19, pp.59-66, as used in Long et al (2020) https://www.frontiersin.org/articles/10.3389/fmars.2020.00460/full
}
\examples{
x <- ImageWidth(Height=0.55, Angle=28.8, VFOV=40.3, HFOV=66.4, Position=0.5)
}
