% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaserValidation.R
\name{ImageWidthFromLasers}
\alias{ImageWidthFromLasers}
\title{Estimate width of image based on laser annotations}
\usage{
ImageWidthFromLasers(LaserXY1, LaserXY2, LaserDistanceM, PixelDimensions)
}
\arguments{
\item{LaserXY1}{\itemize{
\item pixel coordinates (top left is 0,0) of first laser dot
}}

\item{LaserXY2}{\itemize{
\item pixel coordinates (top left is 0,0) of second laser dot
}}

\item{LaserDistanceM}{\itemize{
\item Distance in meters between laser dots
}}

\item{PixelDimensions}{\itemize{
\item dimensions of image in pixels
}}
}
\value{
List containing the following elements
LaserXY1 - Input parameter
LaserXY2 - Input parameter
LaserDistanceM - Input parameter
PixelDimensions - Input parameter
LaserHeight - Height of laser dots in image (percentage of image from bottom)
LaserDistanceXPix - Distance between lasers along the x-axis in pixels
LaserDistanceYPct - Variability in laser dot height (Y axis) expressed as a percentage of image pixels
ImageWidth - Inferred image with at the height of the laser dots.
}
\description{
Estimate width of image based on laser annotations
}
\details{
Estimate width of image based on laser dots. Input requires pixel coordinates of two laser dots seen in an image. It assumes you have a horizontal, flat annotated image.
}
\examples{
x <- ImageWidthFromLasers(c(10, 10), c(50, 10), 0.2, c(2000, 1000))
}
