% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeightsInImage.R
\name{InWaterFOV}
\alias{InWaterFOV}
\title{Convert in-air FOV angle to in-water FOV angle}
\usage{
InWaterFOV(FOV, RefractiveIndex = 1.34)
}
\arguments{
\item{FOV}{value in degrees of the field of view}

\item{RefractiveIndex}{value of refractive index of water (default 1.34 - varies by temperature, salinity & pressure)}
}
\value{
A value representing the in-water field of view in degrees
}
\description{
Convert in-air FOV angle to in-water FOV angle
}
\details{
InWaterFOV() convert in-air field of view angle into in-water field of view angle based on refractive index of water
Formula used is 2 x asin( sin( FOV / 2 ) / RI ) where FOV= Field of View and RI = Refractive Index
The default value of refractive index is 1.34, which is a commonly used value representing shallow marine environments.
RI varies based on temperature, salinity and atmopheric pressure (~depth). A value of 1.33 is often used for shallow fresh water, but can go above 1.35 for very deep, saline, cold water.
You can use a value specific to your water properties if you know Temperature, salinity and pressure using rho::n_water()
(https://github.com/AlexCast/rho/)
}
\examples{
FOV.water <- InWaterFOV(100, RefractiveIndex=1.34)
}
