% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layback.R
\name{Layback}
\alias{Layback}
\title{Estimate position of towed camera using the layback method}
\usage{
Layback(StartLong, StartLat, EndLong, EndLat, Depth, WireLength, Offset = 0)
}
\arguments{
\item{StartLong}{\itemize{
\item Longitude (decimal degrees) of the start of the tow
}}

\item{StartLat}{\itemize{
\item Latitude (decimal degrees) of the start of the tow
}}

\item{EndLong}{\itemize{
\item Longitude (decimal degrees) of the end of the tow
}}

\item{EndLat}{\itemize{
\item Latitude (decimal degrees) of the end of the tow
}}

\item{Depth}{\itemize{
\item Depth (m) of the camera on the seabed
}}

\item{WireLength}{\itemize{
\item Length of wire (m) payed out for the towed camera
}}

\item{Offset}{\itemize{
\item Distance (m) from the GPS reciever to the back of the vessel (default 0)
}}
}
\value{
List containing the following calculations used in estimating layback position
StartLong - Input parameter
StartLat - Input parameter
EndLong - Input parameter
EndLat - Input parameter
Depth - Input parameter
WireLength - Input parameter
Offset - Input parameter
Layback.Distance. - Distance of camera behind vessel
Layback.StartLong - Longitude of camera start position
Layback.StartLat - Latitude of camera start position
Layback.EndLong - Longitude of camera end position
Layback.EndLat - Latitude of camera end position
Depth - Input parameter
}
\description{
Estimate position of towed camera using the layback method
}
\details{
Estimate position of towed camera from the vessel position using the layback method. Trigonometrically calculate the distance of camera from vessel using the water depth and amount of wire payed out. This assumes the camera is directly behind the ship (based on inferred direction of travel).
}
\examples{
x <- Layback(StartLong=0, StartLat=0, EndLong=1, EndLat=1, Depth=100, WireLength=120)
}
