% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaserValidation.R
\name{ValidateWithLasers}
\alias{ValidateWithLasers}
\title{Estimate width of image based on laser annotations}
\usage{
ValidateWithLasers(
  LaserXY1,
  LaserXY2,
  LaserDistanceM,
  PixelDimensions,
  Height,
  Angle,
  VFOV,
  HFOV
)
}
\arguments{
\item{LaserXY1}{\itemize{
\item pixel coordinates (top left is 0,0) of first laser dot
}}

\item{LaserXY2}{\itemize{
\item pixel coordinates (top left is 0,0) of second laser dot
}}

\item{LaserDistanceM}{\itemize{
\item Distance in meters between laser dots
}}

\item{PixelDimensions}{\itemize{
\item dimensions of image in pixels
}}

\item{Height}{\itemize{
\item Height of the camera above the seabed (in m)
}}

\item{Angle}{\itemize{
\item Angle (in dgrees) of the camera pointing down from the horizontal
}}

\item{VFOV}{\itemize{
\item Vertical field of view angle (in degrees)
}}

\item{HFOV}{\itemize{
\item Horizontal field of view angle (in degrees)
}}
}
\value{
List containing the following elements
LaserXY1 - Input parameter
LaserXY2 - Input parameter
LaserDistanceM - Input parameter
PixelDimensions - Input parameter
Height - Input parameter
Angle - Input parameter
VFOV - Input parameter
HFOV - Input parameter
LaserHeight - Height of laser dots in image (proportion of image height 0=bottom, 1=top)
LaserWidth - Laser-based width of image at height of lasers
TrigWidth - Trigometry-based width of image at height of lasers
WidthDifference - Absolute difference between two image width measurements
BestAngle - Camera angle that produces trigonometric image-width calculation that matches laser width (assuming fixed camera height)
BestHeight - Camera height that produces trigonometric image-width calculation that matches laser width (assuming fixed camera angle)
}
\description{
Estimate width of image based on laser annotations
}
\details{
Compare image width calculated with trigonometry and in-image parallel lasers. Input requires pixel coordinates of two laser points seen in an image and image details required for trigonometric area calculation. This assumes a horizontal camera at a fixed oblique angle.
}
\examples{
x <- ValidateWithLasers(c(400, 900), c(700, 900), 0.2, c(2000, 1000), Height=0.55, Angle=28.8, VFOV=40.3, HFOV=66.4)
}
