% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedIntervalStills.R
\name{FixedIntervalStills}
\alias{FixedIntervalStills}
\title{Fetch stills from video at fixed intervals}
\usage{
FixedIntervalStills(
  video,
  outdir = NA,
  window = 0.5,
  interval = 30,
  dark.max = 0.1,
  time.offset = 0
)
}
\arguments{
\item{video}{\itemize{
\item file name of your video survey
}}

\item{outdir}{\itemize{
\item directory for output images (default = where video is located)
}}

\item{window}{\itemize{
\item time window (in seconds) around interval to look for best image (default 1 second)
}}

\item{interval}{\itemize{
\item interval in seconds for stills (default 30 seconds)
}}

\item{dark.max}{\itemize{
\item maximum proportion of dark pixels acceptable for image (default 0.1)
}}

\item{time.offset}{\itemize{
\item start still gathering from this point in the video (time in seconds)
}}
}
\value{
Dataframe of timecode, image files, focus score & darkness pixels
}
\description{
Fetch stills from video at fixed intervals
}
\details{
Fetch best-focus still images from video at fixed intervals
}
\examples{
# FixedIntervalStills("myvideo.mp4", window=0.5, interval=30)
}
